pyproject.toml
[project]
name = "job-hunter"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "python-jobspy>=1.1.79",
    "pandas>=2.2.0",
    "pydantic>=2.6.0",
    "openpyxl>=3.1.2",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "playwright>=1.44.0",
    "feedparser>=6.0.11",
    "rich>=13.7.3",
    "click>=8.1.7",
    "Jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "tenacity>=8.3.0",
    "schedule>=1.2.2",
]

[tool.uv]
dev-dependencies = []
